Refactored whole laser logic
Instead of just using laser sprites, I now use 2 sprites for a laser:

- the LaserHead represents a 5x5 sprite that acts as the "head" of the laser,
  which is always in front of the laser and calculates the moving. It is also
  involved in colissions with other heads, lasers and blocks
- the Laser (tail) is a growing sprite that grows while the head moves. For each
  direction switch, a new laser (tail) sprite is created

Only the head triggers colissions, while tails/tails colissions do nothing.

The only drawback of this is that is _may_ happen that the head moves "over" a laser
tail, when it moves too fast (e.g. too many frames/s). We may see if this could happen...